# api/openapi.yaml
openapi: 3.0.3
info:
  title: FEV-Core API
  description: |
    API de emisión de documentos electrónicos para Colombia.

    Recibe los datos de una operación comercial y produce un documento electrónico
    generado, firmado y transmitido para validación, con su estado consultable en
    todo momento.

    ## Modelo de emisión

    La emisión es asíncrona. La API recibe el documento, le asigna consecutivo y
    responde de inmediato sin contactar al servicio de validación. El procesamiento
    ocurre en segundo plano y el resultado se consulta con `GET /documentos/{id}`.

    Razón: la validación ante la autoridad tributaria tarda varios segundos y puede
    fallar. Ver `docs/adr/0005-emision-asincrona.md`.

    ## Reintentos seguros

    Toda solicitud de emisión incluye `referenciaExterna`, un identificador que el
    integrador asigna a la operación. Si una solicitud se repite con la misma
    referencia, la API devuelve el documento ya creado en lugar de crear otro.

    Esto permite reintentar sin riesgo cuando se pierde la respuesta:

    - Documento creado en esta solicitud: `202 Accepted`
    - Documento ya existía: `200 OK`

    ## Autenticación

    Todas las operaciones requieren una llave de API en la cabecera `X-Api-Key`.

    ## Limitación conocida

    Esta versión opera contra un simulador del servicio de validación, no contra el
    servicio real de la DIAN. Ver `docs/adr/0007-abstraccion-proveedor-validacion.md`.
  version: 1.0.0
  contact:
...
  license:
...
servers:
...
security:
  - LlaveApi: []
paths:
...
  /documentos:
    get:
      tags: [Consulta]
      summary: Listar documentos
      description: Implementa RF-24.
      operationId: listarDocumentos
      parameters:
        - name: tipo
          in: query
          schema:
            $ref: '#/components/schemas/TipoDocumento'
        - name: estado
          in: query
          schema:
            $ref: '#/components/schemas/EstadoDocumento'
        - name: desde
          in: query
          description: Fecha de emisión mínima, inclusive.
          schema:
            type: string
            format: date
        - name: hasta
          in: query
          description: Fecha de emisión máxima, inclusive.
          schema:
            type: string
            format: date
        - name: pagina
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: tamanoPagina
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Página de resultados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginaDocumentos'
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /documentos/{id}:
    get:
      tags: [Consulta]
      summary: Consultar un documento
      description: |
        Devuelve el documento con su estado actual. Si fue aprobado incluye el
        código único; si fue rechazado incluye los errores de validación.

        Implementa RF-22.
      operationId: consultarDocumento
      parameters:
        - $ref: '#/components/parameters/DocumentoId'
      responses:
        '200':
          description: El documento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Documento'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
  /documentos/{id}/historial:
    get:
      tags: [Consulta]
      summary: Consultar el historial de estados
      description: |
        Devuelve todas las transiciones de estado del documento en orden
        cronológico, con la marca de tiempo y el motivo de cada una.

        Implementa RF-23.
      operationId: consultarHistorial
      parameters:
        - $ref: '#/components/parameters/DocumentoId'
      responses:
        '200':
          description: Historial completo.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransicionEstado'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
  /documentos/{id}/xml:
    get:
      tags: [Consulta]
      summary: Descargar el XML
      description: |
        Devuelve el archivo XML tal como se guardó, byte a byte.

        Si el documento ya está firmado, devuelve el **firmado**: es el que
        circula y el que la autoridad recibiría. El XML sin firmar se conserva
        internamente para poder reproducir el cálculo de la firma, no para
        entregarlo.

        Responde `409` con `XML_NO_DISPONIBLE` si aún no se ha generado.

        Implementa RF-25.
      operationId: descargarXml
      parameters:
        - $ref: '#/components/parameters/DocumentoId'
      responses:
        '200':
          description: El XML firmado.
          content:
            application/xml:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '409':
          $ref: '#/components/responses/ConflictoNegocio'

  # ─────────────────────────────  ADQUIRENTES  ─────────────────────────────
...
components:
  securitySchemes:
    LlaveApi:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Llave asignada al sistema integrador. Se muestra una sola vez al crearse y
        no puede recuperarse después (INV-INT-01).
  parameters:
    DocumentoId:
      name: id
      in: path
      required: true
      description: Identificador interno del documento.
      schema:
        type: string
        format: uuid
...
  responses:
...
    ConflictoNegocio:
      description: |
        La solicitud es válida en forma, pero viola una regla de negocio o el estado
        actual del sistema no permite la operación. El campo `codigo` identifica la
        causa. Ver el catálogo en `docs/05-contrato-api.md`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problema'
          examples:
            rangoAgotado:
              summary: No quedan números disponibles
              value:
                type: "https://github.com/Breiner1412/FEV-Core/errors/rango-agotado"
                title: "El rango de numeración está agotado"
                status: 409
                codigo: "RANGO_AGOTADO"
                detail: "El rango SETP con prefijo SETP no tiene números disponibles. Último asignado: 990000999."
                instance: "/api/v1/facturas"
            notaExcede:
              summary: La nota supera el valor de la factura
              value:
                type: "https://github.com/Breiner1412/FEV-Core/errors/nota-excede-factura"
                title: "La nota crédito supera el valor de la factura"
                status: 409
                codigo: "NOTA_EXCEDE_VALOR_FACTURA"
                detail: "El total acumulado de notas crédito (450000.00) superaría el total de la factura (400000.00)."
                instance: "/api/v1/notas-credito"
    NoAutenticado:
      description: Falta la llave de API, o la llave no existe o está desactivada.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problema'
          example:
            type: "https://github.com/Breiner1412/FEV-Core/errors/no-autenticado"
            title: "Credencial inválida"
            status: 401
            codigo: "LLAVE_INVALIDA"
    NoEncontrado:
      description: El recurso solicitado no existe.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problema'
  schemas:
...
    TipoDocumento:
      type: string
      enum: [FACTURA, NOTA_CREDITO, NOTA_DEBITO]
    EstadoDocumento:
      type: string
      description: |
        Estados posibles. Ver la máquina de estados en la sección 6 de
        `docs/02-requerimientos.md`.

        - `RECIBIDO`: aceptado y con consecutivo asignado, pendiente de procesar.
        - `EN_PROCESO`: generando o firmando el XML.
        - `TRANSMITIDO`: entregado al servicio de validación, esperando veredicto.
        - `APROBADO`: validado. Estado terminal.
        - `RECHAZADO`: rechazado por la autoridad. Estado terminal.
        - `FALLIDO`: sin desenlace; hace falta una persona. No implica que la
          autoridad no lo haya recibido (RN-13). El detalle de su transición en
          el historial dice qué consta y si hay que verificar ante la autoridad
          antes de reemplazarlo (ADR-0015). Estado terminal.
      enum: [RECIBIDO, EN_PROCESO, TRANSMITIDO, APROBADO, RECHAZADO, FALLIDO]
    TipoImpuesto:
      type: string
      enum: [IVA, INC, IVA_EXENTO, IVA_EXCLUIDO]
...
    MotivoNota:
      type: string
      enum:
        - DEVOLUCION_PARCIAL
        - ANULACION
        - REBAJA
        - DESCUENTO
        - AJUSTE_PRECIO
        - OTROS

    # ── Solicitudes de emisión ──
...
    Documento:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tipo:
          $ref: '#/components/schemas/TipoDocumento'
        estado:
          $ref: '#/components/schemas/EstadoDocumento'
        referenciaExterna:
          type: string
        prefijo:
          type: string
          example: "SETP"
        consecutivo:
          type: integer
          format: int64
          example: 990000123
        numeroCompleto:
          type: string
          description: Prefijo y consecutivo concatenados, como aparece en el documento.
          example: "SETP990000123"
        fechaEmision:
          type: string
          format: date-time
        adquirenteId:
          type: string
          format: uuid
          description: El adquirente en el catálogo, por si hace falta consultarlo.
        emisor:
          $ref: '#/components/schemas/DatosTributarios'
        adquirente:
          $ref: '#/components/schemas/DatosTributarios'
        lineas:
          type: array
          items:
            $ref: '#/components/schemas/Linea'
        totales:
          $ref: '#/components/schemas/Totales'
        moneda:
          type: string
          example: "COP"
        documentoReferenciadoId:
          type: string
          format: uuid
          nullable: true
        motivo:
          allOf:
            - $ref: '#/components/schemas/MotivoNota'
          nullable: true
          description: Por qué se emitió la nota. Nulo en facturas.
        observaciones:
          type: string
          nullable: true
          maxLength: 500
        codigoUnico:
          type: string
          nullable: true
          description: |
            CUFE en facturas, CUDE en notas. Lo calcula el emisor con SHA-384
            sobre los valores tal como quedan escritos en el XML, y viaja dentro
            de él. Nulo hasta que el XML se genera (RF-20).
        identificadorSeguimiento:
          type: string
          nullable: true
          description: |
            Con qué conoce la autoridad este documento. Nulo hasta que se
            transmite (RF-18).
        erroresValidacion:
          type: array
          description: |
            Por qué se rechazó el documento (RF-21). Lista vacía si no fue
            rechazado; nunca ausente.

            Un documento rechazado no se retransmite: se corrige y se emite uno
            nuevo (RN-07), así que esto es lo único que el integrador puede usar
            para saber qué arreglar.
          items:
            type: string
          example:
            - "FAD06: NIT del emisor no autorizado."
    DocumentoResumen:
      type: object
      description: Versión reducida para listados.
      properties:
        id:
          type: string
          format: uuid
        tipo:
          $ref: '#/components/schemas/TipoDocumento'
        estado:
          $ref: '#/components/schemas/EstadoDocumento'
        numeroCompleto:
          type: string
        fechaEmision:
          type: string
          format: date-time
        adquirenteRazonSocial:
          type: string
        totalAPagar:
          type: number
          format: double
        codigoUnico:
          type: string
          nullable: true
    Linea:
      type: object
      description: |
        Los valores son copias tomadas al momento de emitir. No cambian si el
        producto de origen se modifica después (INV-LIN-03).
      properties:
        numero:
          type: integer
        productoId:
          type: string
          format: uuid
          description: Referencia informativa al producto de origen (INV-LIN-04).
        codigo:
          type: string
        descripcion:
          type: string
        unidadMedida:
          type: string
        cantidad:
          type: number
          format: double
        precioUnitario:
          type: number
          format: double
        descuento:
          type: number
          format: double
        baseGravable:
          type: number
          format: double
        impuestos:
          type: array
          items:
            $ref: '#/components/schemas/ImpuestoLinea'
        total:
          type: number
          format: double
    ImpuestoLinea:
      type: object
      properties:
        tipo:
          $ref: '#/components/schemas/TipoImpuesto'
        tarifa:
          type: number
          format: double
          example: 19.0
        baseGravable:
          type: number
          format: double
        valor:
          type: number
          format: double
    Totales:
      type: object
      description: |
        Calculados siempre a partir de las líneas. El redondeo se aplica en este
        nivel, sobre el total, no línea por línea (RN-06, INV-TOT-02).
      properties:
        totalBrutoAntesImpuestos:
          type: number
          format: double
        totalDescuentos:
          type: number
          format: double
        totalBaseImponible:
          type: number
          format: double
        totalImpuestos:
          type: number
          format: double
        totalAPagar:
          type: number
          format: double
    TransicionEstado:
      type: object
      properties:
        secuencia:
          type: integer
          description: |
            Posición en el historial, empezando en 1. El historial se ordena por
            este campo y no por `ocurridaEn`: dos transiciones pueden compartir
            marca de tiempo.
        estadoAnterior:
          allOf:
            - $ref: '#/components/schemas/EstadoDocumento'
          nullable: true
        estadoNuevo:
          $ref: '#/components/schemas/EstadoDocumento'
        ocurridaEn:
          type: string
          format: date-time
        motivo:
          type: string
          example: "Transmisión aceptada por el servicio de validación."
        detalle:
          type: string
          nullable: true
    DatosTributarios:
      type: object
      description: Copia de los datos vigentes al momento de emitir.
      properties:
        tipoIdentificacion:
          type: string
        identificacion:
          type: string
        digitoVerificacion:
          type: string
          nullable: true
        razonSocial:
          type: string
        direccion:
          type: string
        municipioCodigo:
          type: string
        municipioNombre:
          type: string
        correo:
          type: string
          format: email
        regimen:
          type: string
        responsabilidades:
          type: array
          items:
            type: string

    # ── Catálogos ──
...
    Paginacion:
      type: object
      properties:
        pagina:
          type: integer
        tamanoPagina:
          type: integer
        totalElementos:
          type: integer
          format: int64
        totalPaginas:
          type: integer
    PaginaDocumentos:
      allOf:
        - $ref: '#/components/schemas/Paginacion'
        - type: object
          properties:
            elementos:
              type: array
              items:
                $ref: '#/components/schemas/DocumentoResumen'
...
    Problema:
      type: object
      description: |
        Formato estándar de error (RFC 7807), extendido con `codigo` para que el
        integrador pueda decidir según la causa sin leer el texto.
      properties:
        type:
          type: string
          format: uri
          description: Dirección que describe este tipo de error.
        title:
          type: string
          description: Descripción corta y legible del problema.
        status:
          type: integer
          description: Código de estado HTTP.
        detail:
          type: string
          description: Explicación específica de esta ocurrencia.
        instance:
          type: string
          description: Ruta donde ocurrió.
        codigo:
          type: string
          description: |
            Código estable de la causa. Es lo que el integrador debe usar para
            decidir; `title` y `detail` pueden cambiar de redacción.
          example: "RANGO_AGOTADO"
        traceId:
          type: string
          description: Identificador de correlación para diagnóstico (RNF-10).
...